I/KailuaGame

+++ packages/contracts-bedrock/interfaces/dispute/zk/IKailuaTournament.sol
@@ -1,17 +1,3 @@
-// Copyright 2024, 2025 RISC Zero, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity 0.8.24;
 
@@ -55,7 +41,7 @@ interface IKailuaTournament is IDisputeGame, ISemver {
     function gameData() external view returns (GameType gameType_, Claim rootClaim_, bytes memory extraData_);
     function gameIndex() external view returns (uint256);
     function gameType() external view returns (GameType gameType_);
-    function getChallengerDuration(uint256 asOfTimestamp) external view returns (Duration duration_);
+    function getChallengerDuration(uint64 asOfTimestamp) external view returns (Duration duration_);
     function initialize() external payable;
     function isViableSignature(bytes32 childSignature) external view returns (bool isViableSignature_);
     function l1Head() external pure returns (Hash l1Head_);
