+++ packages/contracts-bedrock/src/dispute/zk/KailuaTournament.sol
@@ -228,7 +228,7 @@ abstract contract KailuaTournament is Clone, IDisputeGame, ISemver {
     }
 
     /// @notice Returns the amount of time left for challenges as of the input timestamp.
-    function getChallengerDuration(uint256 asOfTimestamp) public view virtual returns (Duration duration_);
+    function getChallengerDuration(uint64 asOfTimestamp) public view virtual returns (Duration duration_);
 
     /// @notice Returns the earliest time at which this proposal could have been created
     function minCreationTime() public view virtual returns (Timestamp minCreationTime_);
